[ARCH-P4] Move poisoning evaluation behind application/query
